requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
numpy==2.4.0
portalocker==3.2.0
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.5.0
python-docx==1.2.0
python-dotenv==1.2.1
qdrant-client==1.16.2
requests==2.32.5
soupsieve==2.8.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
